Moving survey items into their own folder
