+++ webpack/landing.js
@@ -13,9 +13,9 @@ import FormShowContainer from './containers/forms/FormShowContainer';
 import QuestionShowContainer from './containers/questions/QuestionShowContainer';
 import ResponseSetEditContainer from './containers/response_sets/ResponseSetEditContainer';
 import QuestionEditContainer from './containers/questions/QuestionEditContainer';
-import SurveyEditContainer from './containers/SurveyEditContainer';
+import SurveyEditContainer from './containers/surveys/SurveyEditContainer';
 import FormsEditContainer from './containers/forms/FormsEditContainer';
-import SurveyShowContainer from './containers/surveys/ShowContainer';
+import SurveyShowContainer from './containers/surveys/SurveyShowContainer';
 import Privacy from './containers/Privacy';
 import Help from './containers/Help';
 import App from './containers/App';
